[Refactor] #147 - 네트워크 코드 이동
